+++ assets/javascripts/kitten/components/molecules/list/index.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 import deprecated from 'prop-types-extra/lib/deprecated'
-import styled, { css } from 'styled-components'
+import styled from 'styled-components'
 import { pxToRem } from '../../../helpers/utils/typography'
 import { ButtonItem } from './components/button-item'
 import classNames from 'classnames'
@@ -11,19 +11,22 @@ const StyledList = styled.ul`
   margin: 0;
   padding: 0;
 
-  ${({ radiusValue }) =>
-    radiusValue > 0 &&
-    css`
-      & li:last-child .k-List__button {
-        border-bottom-left-radius: ${pxToRem(radiusValue)};
-        border-bottom-right-radius: ${pxToRem(radiusValue)};
-      }
-    `}
+  &.k-List--hasBottomRadius .k-List__item:last-child .k-List__button {
+    border-bottom-left-radius: var(--list-bottomRadiusValue, 0);
+    border-bottom-right-radius: var(--list-bottomRadiusValue, 0);
+  }
 
   .k-List__item {
     display: flex;
     gap: ${pxToRem(10)};
     align-items: center;
+
+    &:not(:first-child) {
+      margin-bottom: var(--list-gap);
+    }
+    &:not(:last-child) {
+      margin-bottom: var(--list-gap);
+    }
   }
 
   .k-List__marker {
@@ -41,6 +44,7 @@ export const List = ({
   bottomBorderRadiusValue,
   withBottomBorderRadius,
   markerElement,
+  gap,
   ...props
 }) => {
   const radiusValue =
@@ -51,9 +55,15 @@ export const List = ({
   return (
     <StyledList
       {...props}
-      className={classNames('k-List', className)}
-      styles={style}
-      radiusValue={radiusValue}
+      className={classNames('k-List', className, {
+        'k-List--hasBottomRadius':
+          bottomBorderRadiusValue > 0 || withBottomBorderRadius,
+      })}
+      style={{
+        '--list-gap': pxToRem(gap),
+        '--list-bottomRadiusValue': pxToRem(radiusValue || 0),
+        ...style,
+      }}
     >
       {React.Children.map(children, (child, index) => {
         if (!React.isValidElement(child)) return null
@@ -61,7 +71,9 @@ export const List = ({
         return (
           <li key={index} className="k-List__item">
             {markerElement && (
-              <span aria-hidden className="k-List__marker">{markerElement}</span>
+              <span aria-hidden className="k-List__marker">
+                {markerElement}
+              </span>
             )}
             {child}
           </li>
@@ -76,11 +88,13 @@ List.ButtonItem = ButtonItem
 List.defaultProps = {
   markerElement: null,
   bottomBorderRadiusValue: 0,
+  gap: 0,
 }
 
 List.propTypes = {
   markerElement: PropTypes.node,
   bottomBorderRadiusValue: PropTypes.number,
+  gap: PropTypes.number,
   withBottomBorderRadius: deprecated(
     PropTypes.number,
     'This prop is deprecated, please use `bottomBorderRadiusValue instead.',